Couple changes to model locations
Updated some file locations and project directory changes. Made more things actual model classes and moved some models from the ViewSupport to the models folder where needed
