+++ FulcrumInjector/FulcrumInjector.csproj
@@ -172,17 +172,14 @@
     </ApplicationDefinition>
     <Compile Include="FulcrumViewContent\FulcrumModels\LogFileModels\FulcrumLogFileModel.cs" />
     <Compile Include="FulcrumViewContent\FulcrumModels\LogFileModels\FulcrumLogFileSet.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogConfiguration.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogEvents\FileAccessedEventArgs.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogEvents\FileEventArgs.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogEvents\FileModifiedEventArgs.cs" />
+    <Compile Include="FulcrumViewContent\FulcrumModels\WatchdogModels\WatchdogConfiguration.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumDataConverters\EnumLibExtensions.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumDataConverters\SizeFormatExtensions.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumDataConverters\TextTableGenerator.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumFrameAnimator.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumStyles\AppStyleModels\AppStyleEnums.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumUpdater\FulcrumUpdater.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumUpdater\FulcrumVersionInfo.cs" />
+    <Compile Include="FulcrumViewSupport\FulcrumUpdater.cs" />
+    <Compile Include="FulcrumViewContent\FulcrumModels\FulcrumVersionInfo.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumEmailBroker.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumNavMenuItem.cs" />
     <Compile Include="FulcrumViewContent\FulcrumModels\SettingsModels\FulcrumSettingsShare.cs" />
@@ -244,9 +241,9 @@
     <Compile Include="FulcrumViewSupport\FulcrumDataConverters\BoolToVisibilityConverter.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumDataConverters\HamburgerNavService.cs" />
     <Compile Include="FulcrumViewSupport\FulcrumStyles\CustomColorShader.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogFolder.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogFile.cs" />
-    <Compile Include="FulcrumViewSupport\FulcrumWatchdog\WatchdogService.cs">
+    <Compile Include="FulcrumViewContent\FulcrumModels\WatchdogModels\WatchdogFolder.cs" />
+    <Compile Include="FulcrumViewContent\FulcrumModels\WatchdogModels\WatchdogFile.cs" />
+    <Compile Include="FulcrumViewSupport\FulcrumWatchdogService.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="FulcrumViewSupport\FulcrumWin32Invokers.cs" />
