+++ FulcrumInjector/Properties/AssemblyInfo.cs
@@ -19,7 +19,7 @@ using System.Resources;
 [assembly: Guid("8cb7e832-9e90-4820-b225-0a4d59e6c0a2")]
 
 // Version information
-[assembly: AssemblyVersion("6.3.24.2348")]
-[assembly: AssemblyFileVersion("6.3.24.2348")]
+[assembly: AssemblyVersion("6.3.25.2349")]
+[assembly: AssemblyFileVersion("6.3.25.2349")]
 [assembly: NeutralResourcesLanguageAttribute( "en-US" )]
